+++ src/Sentry/Laravel/ServiceProvider.php
@@ -54,8 +54,8 @@ class ServiceProvider extends BaseServiceProvider
         Features\CacheIntegration::class,
         Features\QueueIntegration::class,
         Features\ConsoleIntegration::class,
+        Features\Storage\Integration::class,
         Features\LivewirePackageIntegration::class,
-        Features\StorageIntegration::class,
     ];
 
     /**
